third commit

+++ Shop.Client/Client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { AfterViewInit, Component, ElementRef, Input, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 import { NavigationExtras, Router } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
@@ -7,19 +7,61 @@ import { IBasket } from 'src/app/shared/models/basket';
 import { IOrder } from 'src/app/shared/models/order';
 import { CheckoutService } from '../checkout.service';
 
+declare var Stripe: (arg0: string) => any;
+
 @Component({
   selector: 'app-checkout-payment',
   templateUrl: './checkout-payment.component.html',
   styleUrls: ['./checkout-payment.component.scss']
 })
-export class CheckoutPaymentComponent implements OnInit {
+export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
 
   @Input() checkoutForm : FormGroup;
+  @ViewChild('cardNumber', {static : true}) cardNumberElement : ElementRef;
+  @ViewChild('cardExpiry', {static : true}) cardExpiryElement : ElementRef;
+  @ViewChild('cardCvc', {static : true}) cardCvcElement : ElementRef;
+
+  //this is a js lib
+  stripe : any;
+  cardNumber : any;
+  cardExpiry : any;
+  cardCvc : any;
+  cardErrors : any;
+  cardHandler = this.onChange.bind(this);
 
   constructor(private basketService : BasketService, private checkoutService : CheckoutService,
     private toastr : ToastrService, private router : Router) { }
 
-  ngOnInit(): void {
+  ngAfterViewInit(): void {
+    this.stripe = Stripe('pk_test_51M3zFVBRpet1bJMDo3VpSgXusrkS7WhvV51i3lulb9FskXhkLgpIrL3J0TV65sesf5ai90kAfvIBlTHTSxisgLj500NPyDYuXK');
+    const elements = this.stripe.elements();
+
+    this.cardNumber = elements.create('cardNumber');
+    this.cardNumber.mount(this.cardNumberElement.nativeElement);
+    this.cardNumber.addEventListener('change', this.cardHandler);
+
+    this.cardExpiry = elements.create('cardExpiry');
+    this.cardExpiry.mount(this.cardExpiryElement.nativeElement);
+    this.cardExpiry.addEventListener('change', this.cardHandler);
+
+    this.cardCvc = elements.create('cardCvc');
+    this.cardCvc.mount(this.cardCvcElement.nativeElement);
+    this.cardCvc.addEventListener('change', this.cardHandler);
+  }
+
+  ngOnDestroy(): void {
+      this.cardNumber.destroy();
+      this.cardExpiry.destroy();
+      this.cardCvc.destroy();
+  }
+
+  onChange({error} : any) : any{
+    if(error){
+      this.cardErrors = error.message;
+    }
+    else{
+      this.cardErrors = null;
+    }
   }
 
   submitOrder(){
@@ -28,9 +70,24 @@ export class CheckoutPaymentComponent implements OnInit {
 
     this.checkoutService.createOrder(orderToCreate).subscribe((order : any) =>{
       this.toastr.success('Order Created Successfully');
-      this.basketService.deleteLocalBasket(basket.id);
-      const navExtras : NavigationExtras = {state : order};
-      this.router.navigate(['checkout/success'], navExtras);
+      this.stripe.confirmCardPayment(basket.clientSecret, {
+        payment_method : {
+          card : this.cardNumber,
+          billing_details : {
+            name : this.checkoutForm.get('paymentForm').get('nameOnCard').value
+          }
+        }
+      }).then((result : any) => {
+        console.log(result);
+        if(result.paymentIntent){
+           this.basketService.deleteLocalBasket(basket.id);
+           const navExtras : NavigationExtras = {state : order};
+           this.router.navigate(['checkout/success'], navExtras);
+        }else{
+          this.toastr.error('Payment Failed');
+        }
+      });
+
     }, error =>{
       this.toastr.error(error.message);
       console.log(error);
